+++ src/components/home/medicineNote/MedicineList.js
@@ -1,3 +1,4 @@
+// components/home/medicineNote/MedicineList.js
 import React, { useState } from 'react';
 import {
   MediListContainer,
@@ -8,21 +9,11 @@ import {
 } from 'styles/home/MedicineList-styled';
 import Modal from './Modal'; // Modal 컴포넌트 import
 import icon_plus from 'assets/images/icon_plus.png';
+import useMedicineList from 'hooks/home/useMedicineList';
 
 function MedicineList() {
-  const [medi, setMedi] = useState([
-    { id: 1, text: '오메가3', time: null, completed: false },
-    { id: 2, text: '메가슬립 수면엔 미강테아닌', time: null, completed: false },
-  ]);
+  const { medi, handleToggle, addNewMedicine } = useMedicineList();
   const [isModalOpen, setIsModalOpen] = useState(false);
-
-  // 영양제 체크 상태 변경 함수
-  const handleToggle = (id) => {
-    const updatedMedi = medi.map((item) =>
-      item.id === id ? { ...item, completed: !item.completed } : item
-    );
-    setMedi(updatedMedi);
-  };
 
   // 모달 열기 함수
   const openModal = () => {
@@ -32,16 +23,6 @@ function MedicineList() {
   // 모달 닫기 함수
   const closeModal = () => {
     setIsModalOpen(false);
-  };
-
-  // 새로운 영양제 추가 함수
-  const addNewMedicine = (newMedicine, time) => {
-    const nextId = medi.length + 1;
-    const newMedi = [
-      ...medi,
-      { id: nextId, text: newMedicine, time: time, completed: false },
-    ];
-    setMedi(newMedi);
   };
 
   return (
